executeE. Frameworks.
Add executeE that allows us to perform administrative changes to the event network on the fly.
Percolate administrative stuff to the Reactive.Banana.Frameworks module.

FIXME: The trick of using a type class constraint `forall t. Setup t => Moment t a` does not work straight away. It generates an ambiguous type when trying to discharge the Setup constraint.