Correct space semantics for `stepper`.
Any value that is stored in the graph over more than a single atomic evaluation must be stored in WHNF. This implies that latch values must be forced to WHNF before storing them. On the other hand, only latch values can be stored for a longer time, so they are the only place where we need to do this.

The model semantics have been corrected in a similar way.