Pull-based behaviors experiemental implementation
The motivation: when doing DOM programming we sometimes wish to deal
with the dimensions of elements. It seems natural to say that there is a
`Behavior Rect` for an element's bounding client rectangle. But how to
construct such a thing?? We would need an event to give the changes, but
the DOM doesn't supply that, and polling it would probably grow
prohibitively expensive. Why not compute it on-demand?

This patch defines `fromPull :: IO a -> MomentIO (Behavior a)` to
produce pull-based behaviors. Ideally, the IO would be run at most once
per evaluation of the network. In this implementation, it's run at
least 0 times.