Prim.addHandler now takes an Output arg.
The Output "o" is passed down the call stack via
"buildLater" and will eventually be added to the
Network nOutputs ordered bag. It is created here
because we need to return a function that removes it
from the bag, but the result cannot be passed back to
us from the future.