Prevent a deadlock if stepping the network crashes
Currently, the following problem causes a deadlock:

```haskell
{-# language TypeApplications #-}
import Control.Exception
import Data.Functor
import Reactive.Banana
import Reactive.Banana.Frameworks

main :: IO ()
main = do
  (addHandler, fire) <- newAddHandler
  network <- compile $ do
    e <- fromAddHandler addHandler
    e' <- execute $ e $> do
      error "BANG"
      return ()

    reactimate $ print <$> e
    reactimate $ print <$> e'

  actuate network
  _ <- try @SomeException (fire ())

  fire ()
```

The problem occurs when first call `fire`. The high-level step function
first takes the network `MVar`, and then tries to step the network. This
crashes, but the overall program is safe because it was wrapped in
`try`. However, we're now in a state where the network `MVar` is empty,
so the second time we try and call `fire` we deadlock.

The fix in this commit is to bracket the whole network evaluation with
`bracketOnError`, and to restore the starting state if an error occurs.